Added and styled some missing pages, including a splash page, 404, 401, and some misc pages. 404 page is not yet reachable due to django issues... will figure out later

+++ squirespace/blog/views.py
@@ -30,18 +30,21 @@ def registration_complete(request):
     return render_to_response('registration/register_success.html')
 
 def login(request):
+    if not(request.user.is_anonymous()):
+        return render(request, 'registration/already_authenticated.html')
     form = PostForm()
     return render(request, 'blog/login.html', {'form': form})
 
 def post_list(request):
-    posts = Post.objects.filter(published_date__lte=timezone.now())
     if (request.user.is_anonymous()):
-        friends = []
-    else:
-        friends = Friend.objects.friends(request.user)
+        return render(request, 'blog/splash_page.html')
+    posts = Post.objects.filter(published_date__lte=timezone.now())
+    friends = Friend.objects.friends(request.user)
     return render(request, 'blog/post_list.html', {'posts': posts, 'friends': friends})
 
 def post_detail(request, pk):
+    if (request.user.is_anonymous()):
+        return render(request, 'blog/401.html')
     post = get_object_or_404(Post, pk=pk)
     if request.method == "POST":
         form = CommentForm(request.POST)
@@ -56,6 +59,8 @@ def post_detail(request, pk):
     return render(request, 'blog/post_with_comments.html', {'form': form,'post': post})
 
 def post_new(request):
+    if (request.user.is_anonymous()):
+        return render(request, 'blog/401.html')
     if request.method == "POST":
         form = PostForm(request.POST, request.FILES)
         if form.is_valid():
@@ -69,6 +74,8 @@ def post_new(request):
     return render(request, 'blog/post_edit.html', {'form': form})
 
 def post_edit(request, pk):
+    if (request.user.is_anonymous()):
+        return render(request, 'blog/401.html')
     post = get_object_or_404(Post, pk=pk)
     if request.method == "POST":
         form = PostForm(request.POST, instance=post)
@@ -83,11 +90,15 @@ def post_edit(request, pk):
     return render(request, 'blog/post_edit.html', {'form': form,'post':post,'edit':True})
 
 def post_delete(request, pk):
+    if (request.user.is_anonymous()):
+        return render(request, 'blog/401.html')
     post = get_object_or_404(Post, pk=pk)
     post.delete()
     return redirect('post_list')
 
 def profile(request,pk):
+    if (request.user.is_anonymous()):
+        return render(request, 'blog/401.html')
     profile_owner = User.objects.get(id=pk)
     posts = Post.objects.filter(author=profile_owner,published_date__lte=timezone.now()).order_by('published_date')
     
@@ -98,31 +109,46 @@ def profile(request,pk):
     return render(request, 'blog/profile.html', {'user': request.user, 'profile_owner': profile_owner, 'posts': posts, 'friends': friends, 'following':following, 'followers':followers, 'friend_requests':friend_requests})
 
 def send_friend_request(request, pk):
+    if (request.user.is_anonymous()):
+        return render(request, 'blog/401.html')
+    # NEED TO CHECK IF FRIEND IS ANONYMOUSUSER (FRIEND.IS_ANONYMOUS())
     other_user = User.objects.get(id = pk)
     Friend.objects.add_friend(request.user, other_user, message = 'I would like to request your friendship.')
     return render(request, 'blog/profile.html', {'user': request.user, 'other_user': other_user, 'posts': posts})
 
 def accept_friend_request(request, pk):
+    if (request.user.is_anonymous()):
+        return render(request, 'blog/401.html')
+    # NEED TO CHECK IF FRIEND IS ANONYMOUSUSER (FRIEND.IS_ANONYMOUS())
     friend_request = FriendshipRequest.objects.get_object_or_404(pk = pk)
     friend_request.accept_friend_request()
     return render(request, 'blog/profile.html', {'user': request.user})
 
 def reject_friend_request(request, pk):
+    if (request.user.is_anonymous()):
+        return render(request, 'blog/401.html')
+    # NEED TO CHECK IF FRIEND IS ANONYMOUSUSER (FRIEND.IS_ANONYMOUS())
     friend_request = FriendshipRequest.objects.get_object_or_404(pk = pk)
     friend_request.reject_friend_request()
     return render(request, 'blog/profile.html', {'user': request.user})
 
 def cancel_friend_request(request, pk):
+    if (request.user.is_anonymous()):
+        return render(request, 'blog/401.html')
+    # NEED TO CHECK IF FRIEND IS ANONYMOUSUSER (FRIEND.IS_ANONYMOUS())
     friend_request = FriendshipRequest.objects.get_object_or_404(pk = pk)
     friend_request.cancel_friend_request()
     return render(request, 'blog/profile.html', {'user': request.user})
 
 def show_friends(request, pk):
+    if (request.user.is_anonymous()):
+        return render(request, 'blog/401.html')
     if request.method =="GET":
         return Friend.objects.friends(request.user)
 
 
 
-
+def page_not_found(request):
+    return render(request, 'blog/404.html')
